Import units
Also changed getter/setter implementation for Objects, although it means
they can't be displayed in the editor window (this is bad).
Might revert to the old, less clean but more functional way, or else
need to implement my own Editor UI for Synctory. As in-game UI will be
necessary anyway, this might not be the worst option.

NEXT: Load entities and steps, and finish SynctoryHelpers to get them
from their keys